--- public/lib/queries/select.ts
@@ -1,6 +1,0 @@
-import { db } from "@/lib/db";
-import { user } from "@/lib/schema";
-export async function getUsers() {
-    const users = await db.select().from(user)
-    return users;
-}
