+++ src/app/(anti-protected)/onboarding/page.tsx
