+++ src/app/(anti-protected)/page.tsx
@@ -1,6 +1,6 @@
-// import { Button } from "@/components/ui/button"
-import SignIn from "./GoogleSignin"
+import SignIn from "@/components/login/GoogleSignin"
 
+// import { Button } from "@/components/ui/button"
 export default function Home() {
   return (
     <div>
